@@ -1,6 +1,6 @@
 package com.hilfritz.mvp.framework;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 /**IMPORTANT: FRAMEWORK CLASS**/
 
 /**
